remove exported custom styles into a MUIStyles.js
