+++ src/components/ConfirmDialog.jsx
@@ -8,7 +8,7 @@ import {
 	Button,
 } from '@mui/material';
 import CloseIcon from '@mui/icons-material/Close';
-import { buttonStyle } from '../views';
+import { buttonStyle } from './index';
 import './ConfirmDialog.css';
 
 // MUI's Dialog already comes with built-in focus management and accessibility features.
@@ -21,11 +21,6 @@ const Transition = React.forwardRef(function Transition(props, ref) {
 
 const dialogStyle = {
 	backgroundColor: 'rgb(20, 20, 20)',
-};
-
-export const typographyStyle = {
-	padding: '1em',
-	color: 'white',
 };
 
 export function ConfirmDialog({ props }) {
