+++ src/components/index.js
@@ -6,3 +6,4 @@ export * from './TextInputElement';
 export * from './RadioInputElement';
 export * from './ConfirmDialog';
 export * from './DeleteIconWithTooltip';
+export * from './MUIStyles';
