+++ src/views/List.jsx
@@ -11,8 +11,13 @@ import {
 	Button,
 } from '@mui/material';
 import { ArrowDropDown, ArrowDropUp } from '@mui/icons-material';
-import { ListItem, AddItems, TextInputElement } from '../components';
-import { darkPaperStyle, lightPaperStyle } from '../App';
+import {
+	ListItem,
+	AddItems,
+	TextInputElement,
+	darkPaperStyle,
+	lightPaperStyle,
+} from '../components';
 
 // React.memo is needed to prevent unnecessary re-renders of the List component
 // when the props (data and listPath) haven't changed,
